+++ test/drab_test.exs
@@ -7,6 +7,7 @@ defmodule DrabTest do
     assert Drab.config[:events_to_disable_while_processing] == ["click"]
     assert Drab.config[:disable_controls_when_disconnected] == true
     assert Drab.config[:socket] == "/drab/socket"
+    assert Drab.config[:drab_store_storage] == :session_storage
   end